requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
drf-yasg>=1.20,<2.0
pre-commit>=2.12,<3.0
